style: rename dir
