+++ design-toolkit/components/src/index.ts
@@ -16,9 +16,6 @@
 
 // biome-ignore-all assist/source/organizeImports: This comment is used to prevent the biome tool from altering the import statements in this file.
 
-export { Link } from './components/Link';
-export { LinkStyles } from './components/Link/styles';
-export type { LinkProps } from './components/Link/types';
 export { Accordion, AccordionContext } from './components/accordion';
 export {
   AccordionStyles,
@@ -184,6 +181,9 @@ export type { LabelProps } from './components/label/types';
 export { Lines } from './components/lines';
 export { LinesStyles } from './components/lines/styles';
 export type { LinesProps, LinesStylesVariants } from './components/lines/types';
+export { Link } from './components/link';
+export { LinkStyles } from './components/link/styles';
+export type { LinkProps } from './components/link/types';
 export { Menu, MenuContext } from './components/menu';
 export { MenuStyles, MenuStylesDefaults } from './components/menu/styles';
 export type {
